+++ create_recipe_database.py
@@ -162,5 +162,7 @@ def write_recipes_in_file(recipes):
 
 
 if __name__ == '__main__':
-    cleaned_recipes_list = write_recipes_in_list(5)
-    write_recipes_in_file(cleaned_recipes_list)
+
+    pass
+    # cleaned_recipes_list = write_recipes_in_list(5)
+    # write_recipes_in_file(cleaned_recipes_list)
