+++ recommend_collaborative.py
@@ -5,4 +5,4 @@ def recommend():
     """
     create or update the recommendations.csv file
     """
-    ItemKNN(train_file="ratings.csv", output_file="recommendations.csv", sep=", ").compute()
+    ItemKNN(train_file="ratings.csv", output_file="recommendations_collaborative.csv", sep=", ").compute()
