--- recommendations.csv
@@ -1,2 +1,0 @@
-2, 639535, 0.452172
-2, 664473, 0.452172
