+++ recommendations_collaborative.csv
@@ -1,0 +1,10 @@
+2, 652332, 1.245868
+3, 665029, 0.821680
+3, 652332, 0.693998
+4, 649931, 0.712516
+4, 652332, 0.693998
+4, 665029, 0.000000
+5, 649931, 0.000000
+5, 652332, 0.000000
+5, 660146, 0.000000
+5, 665029, 0.000000
