+++ recommendations_content_based.csv
@@ -1,0 +1,10 @@
+2, 652332, 0.510922
+3, 652332, 0.582305
+3, 665029, 0.000000
+4, 652332, 0.594690
+4, 649931, 0.000000
+4, 665029, 0.000000
+5, 652332, 0.622015
+5, 660146, 0.599518
+5, 649931, 0.000000
+5, 665029, 0.000000
