+++ similarities.csv
@@ -1,4 +1,3 @@
-ID_1, ID_2, Similarity, Title_1, Title_2
 652332, 660146, 0.5673654508948414, Monastery soup, Simple Squash Salad
 652332, 716218, 0.6220151615431392, Monastery soup, Tomato and Eggplant Caponata
 652332, 649931, 0.5575357720739681, Monastery soup, Lentil Salad With Vegetables
